Clasificare/requirements.txt
keras
python-resize-image
progressbar2
matplotlib
h5py
sklearn
scikit-image